pyproject.toml
[tool.poetry]
name = "tms-drf-app-woman"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "tms_drf_app_woman"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.6"
psycopg2-binary = "^2.9.5"
pre-commit = "^3.0.4"
flake8 = "^6.0.0"
drf-yasg = "^1.21.5"
django-silk = "^5.0.3"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
mypy = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
